Don't test doc comments.
https://github.com/rust-lang-nursery/rust-bindgen/pull/1384 makes bindgen
respect whitespace in documentation comments instead of trimming them.

This means that some of the comments from the C++ examples try to get compiled
as Rust, with a bad result for `cargo test`s purposes.

Just test using cargo test --lib. An alternative would be to disable doc
comments in bindgen.