Upgrade mach
This change upgrades the version of mach in the repository to match the
one from the latest commit in the Gecko repository. This should allow
mozjs to build with newer versions of Python. Notably, this change does
not upgrade any other Mozilla python packages such as mozbuild so that
they maintain compatibility with the build files of SpiderMonkey. This
is meant as a stopgap to fix compilation on newer versions of Python
until a proper SpiderMonkey upgrade, which will require a significant
amount of work downstream in Servo.